feat(app): use nanoseconds in vehicle control API (#242)
duration parameter in vehicle slowDown, stop, changeSpeed, and changeLane is now interpreted as nanoseconds and has been changed to type long.

+++ app/tutorials/example-applications/src/main/java/org/eclipse/mosaic/app/examples/commuter/SimpleCommuterApp.java
@@ -71,7 +71,7 @@ public class SimpleCommuterApp extends AbstractApplication<VehicleOperatingSyste
         if (!initialTripPlanned) {
             IRoadPosition roadPosition =
                     RoadPositionFactory.createAtEndOfRoute(getOs().getNavigationModule().getCurrentRoute(), 0);
-            getOs().stop(roadPosition, VehicleStopMode.PARK_ON_ROADSIDE, Integer.MAX_VALUE);
+            getOs().stop(roadPosition, VehicleStopMode.PARK_ON_ROADSIDE, Long.MAX_VALUE);
             initialTripPlanned = true;
             stopWatch.start();
         }
